Completed the projects page frontend with all functionalities, delete functionality is to be added to teams, quests and organization page, up -next

+++ backend/workStackBackend/projects/api/views.py
@@ -29,6 +29,9 @@ class ProjectListView(APIView):
         return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
 
 # add permissions to ensure the user is authorized for the same    
+def UserAccess(user,project):
+    return (user==project.created_by)
+
 class ProjectDetailView(APIView):
     def get(self,request,pk):
         try:
@@ -40,20 +43,35 @@ class ProjectDetailView(APIView):
     def patch(self,request,pk):
         try:
             project=Project.objects.get(pk=pk)
-            serializer=ProjectSerializer(project,data=request.data,partial=True)
-
-            if serializer.is_valid():
-                serializer.save()
-                return Response(serializer.data)
-            return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
+            if UserAccess(request.user,project):
+                serializer=ProjectSerializer(project,data=request.data,partial=True)
+                if serializer.is_valid():
+                    serializer.save()
+                    return Response(serializer.data)
+                return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
+            return Response({"message":"Your access is revoked."},
+                            status=status.HTTP_403_FORBIDDEN)
         except Project.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
         
     def delete(self,request,pk):
         try:
             project=Project.objects.get(pk=pk)
-            project.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+            if request.user==project.created_by:
+                project.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            return Response({"message":"your access is revoked."},
+                            status=status.HTTP_403_FORBIDDEN)
         except Project.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
+        
+class ProjectPermissionView(APIView):
+    def get(self,request,pk):
+        try:
+            project=Project.objects.get(pk=pk)
+            acc=UserAccess(request.user,project)
+            return Response({"access":acc},status=status.HTTP_200_OK)
+        except:
+            return Response({"message":"Project doesnt exist."},status=status.HTTP_404_NOT_FOUND)
+
